+++ RELEASES.md
@@ -106,6 +106,7 @@ This release adds support for Python 3.14 with the following limitations:
 - Fixed Polymarket account balance update timing issue (#3161), thanks for reporting @santivazq
 - Fixed Polymarket handling of overfilled FOK orders using `allow_overfills` execution engine config option (#3221), thanks for reporting @Javdu10
 - Fixed Polymarket `match_time` timestamp parsing (#3273), thanks for reporting @santivazq
+- Fixed Tardis book snapshot to deltas CLEAR prepending
 
 ### Internal Improvements
 - Added BitMEX submit broadcaster